code review fixes : move auxilary functions from pipeline to transformation helper
