+++ src/plugins/intel_gna/src/transformations/utils/transformation_helper.cpp
@@ -6,12 +6,13 @@
 
 #include "log/debug.hpp"
 #include "openvino/core/rt_info.hpp"
-#include "openvino/opsets/opset7.hpp"
+#include "openvino/opsets/opset12.hpp"
 #include "openvino/pass/pattern/op/wrap_type.hpp"
 #include "ops/gna_convolution.hpp"
 #include "ops/gna_max_pool.hpp"
+#include "transformations/rt_info/transpose_sinking_attr.hpp"
 
-using namespace ov::opset7;
+using namespace ov::opset12;
 
 namespace ov {
 namespace intel_gna {
@@ -188,6 +189,67 @@ ov::AxisVector reverse_transpose_order(const ov::AxisVector& axis_order) {
         out.at(axis_order[i]) = i;
     }
     return out;
+}
+
+ov::NodeVector find_input_transposes(const std::shared_ptr<const ov::Node>& node) {
+    ov::NodeVector transposes;
+    for (size_t input_idx = 0; input_idx < node->get_input_size(); ++input_idx) {
+        auto input_node = node->get_input_node_shared_ptr(input_idx);
+        auto transpose_node = ov::as_type_ptr<Transpose>(input_node);
+        if (transpose_node)
+            transposes.push_back(transpose_node);
+    }
+
+    return transposes;
+}
+
+void mark_input_transposes_as_nosinking(std::shared_ptr<const ov::Node> node) {
+    for (const auto& input : find_input_transposes(node))
+        ov::mark_as_no_sinking_node(input);
+}
+
+TransposeInfo get_first_input_transpose(const std::shared_ptr<const ov::Node>& node) {
+    for (size_t input_idx = 0; input_idx < node->get_input_size(); ++input_idx) {
+        std::shared_ptr<Node> input_node = node->get_input_node_shared_ptr(input_idx);
+        auto transpose_node = as_type_ptr<Transpose>(input_node);
+        if (!transpose_node)
+            continue;
+        auto constant_node = as_type_ptr<Constant>(transpose_node->input_value(1).get_node_shared_ptr());
+        if (!constant_node)
+            continue;
+        {
+            TransposeInfo input_info;
+            input_info.transpose = transpose_node;
+            input_info.transpose_const = constant_node;
+            return input_info;
+        }
+    }
+
+    return {};
+}
+
+TransposeInfo get_first_output_transpose(const std::shared_ptr<const ov::Node>& node) {
+    for (size_t i = 0; i < node->get_output_size(); ++i) {
+        for (const auto& input : node->output(i).get_target_inputs()) {
+            Node* node = input.get_node();
+            if (!dynamic_cast<Transpose*>(node))
+                continue;
+            auto transpose_node = as_type_ptr<Transpose>(node->shared_from_this());
+            if (!transpose_node)
+                continue;
+            auto constant_node = as_type_ptr<Constant>(transpose_node->input_value(1).get_node_shared_ptr());
+            if (!constant_node)
+                continue;
+            {
+                TransposeInfo input_info;
+                input_info.transpose = transpose_node;
+                input_info.transpose_const = constant_node;
+                return input_info;
+            }
+        }
+    }
+
+    return {};
 }
 
 }  // namespace helper
